Move usb.{cc,h} into its own directory.
