+++ mkninja.sh
@@ -190,6 +190,7 @@ buildlibrary libbackend.a \
     lib/fluxsource/kryoflux.cc \
     lib/fluxsource/sqlitefluxsource.cc \
     lib/fluxsource/streamfluxsource.cc \
+    lib/usb/usb.cc \
     lib/globals.cc \
     lib/hexdump.cc \
     lib/ldbs.cc \
@@ -197,7 +198,6 @@ buildlibrary libbackend.a \
     lib/sector.cc \
     lib/sectorset.cc \
     lib/sql.cc \
-    lib/usb.cc \
     lib/visualiser.cc \
     lib/writer.cc \
